# packages/web/public/openapi.yaml
openapi: 3.0.3
info:
  title: Royal Battalion API
  description: |
    REST API for the Royal Battalion gaming community management platform.
    Handles user authentication (Discord OAuth), whitelist management,
    ticket/prospect tracking, match history, server administration,
    and Discord bot integration.

    ## Versioning
    All REST endpoints are versioned and served under the `/v1` prefix, which is
    baked into the `servers[].url` base below (so the paths documented here are
    relative to `/v1`, e.g. `/users` resolves to `/v1/users`).

    A few root-level, NON-versioned endpoints are exempt (external / infra
    contracts that must not move) and declare their own root `servers` override:
    - `GET /health` -- infra/Railway health probe.
    - `GET /admins.cfg` -- fetched by the Squad game server at a fixed URL.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained via the `/auth/sync` endpoint.
    The flow is:
    1. User authenticates with Discord OAuth on the frontend
    2. Frontend sends the Discord access token to `POST /auth/sync`
    3. API returns a JWT (valid 4 hours) containing userId and permissions
    4. Include the JWT in all subsequent requests as `Authorization: Bearer <token>`

    ## Permissions
    Access is controlled by a granular RBAC system with 19 permissions.
    Users with the `admin` permission bypass all permission checks.

    ## Rate Limiting
    - Global: 200 requests/minute per IP
...

...
  version: 1.0.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /squadjs-config:
    get:
      tags: [SquadJS Config]
      summary: List available environments
      description: Returns the list of available SquadJS configuration environments.
      responses:
        "200":
          description: Environment list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      environments:
                        type: array
                        items:
                          type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          description: GITHUB_CONFIG_TOKEN not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-required-permissions:
        - view:squadjs
        - manage:squadjs
  /squadjs-config/descriptions:
    get:
      tags: [SquadJS Config]
      summary: Get plugin descriptions
      description: Fetches plugin and field descriptions from SquadJS source code on GitHub.
      responses:
        "200":
          description: Plugin descriptions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      descriptions:
                        type: object
                        additionalProperties:
                          type: string
                      fieldDescriptions:
                        type: object
                        additionalProperties:
                          type: object
                          additionalProperties:
                            type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - view:squadjs
        - manage:squadjs
  /squadjs-config/{env}:
    get:
      tags: [SquadJS Config]
      summary: Read environment config
      description: Reads the SquadJS plugin configuration for a specific environment from GitHub.
      parameters:
        - name: env
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plugin configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      plugins:
                        type: array
                        items:
                          $ref: "#/components/schemas/SquadJSPlugin"
                      environment:
                        type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          description: GitHub token not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-required-permissions:
        - view:squadjs
        - manage:squadjs
    patch:
      tags: [SquadJS Config]
      summary: Write environment config
      description: Partially writes updated plugin configuration to GitHub for a specific environment (merges plugins).
      parameters:
        - name: env
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plugins]
              properties:
                plugins:
                  type: array
                  items:
                    $ref: "#/components/schemas/SquadJSPlugin"
      responses:
        "200":
          description: Config saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      saved:
                        type: boolean
                        example: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          description: GitHub token not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-required-permissions:
        - manage:squadjs

  # ── Server Config ─────────────────────────────────────
  /server-config:
    get:
      tags: [Server Config]
      summary: List server configs
      description: Returns all server configurations. SFTP passwords are masked.
      responses:
        "200":
          description: Server configs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ServerConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - view:whitelist
    post:
      tags: [Server Config]
      summary: Create or update server config
      description: Upserts a server configuration. SFTP password is encrypted if ENCRYPTION_KEY is set.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [server, label]
              properties:
                server:
                  type: string
                  minLength: 1
                label:
                  type: string
                  minLength: 1
                sftpHost:
                  type: string
                  nullable: true
                sftpPort:
                  type: integer
                  default: 22
                sftpUser:
                  type: string
                  nullable: true
                sftpPass:
                  type: string
                  nullable: true
                sftpPath:
                  type: string
                  nullable: true
                syncEnabled:
                  type: boolean
      responses:
        "200":
          description: Config saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/ServerConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - admin
  /server-config/{server}:
    patch:
      tags: [Server Config]
      summary: Set SFTP sync state
      description: |
        Sets the SFTP sync enabled flag for a server to the client-supplied state
        (idempotent). Replaces the old `PUT /server-config/:server/sync` blind toggle.
      parameters:
        - name: server
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [syncEnabled]
              properties:
                syncEnabled:
                  type: boolean
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/ServerConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Server config not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-required-permissions:
        - manage:whitelist-sync
    delete:
      tags: [Server Config]
      summary: Delete server config
      parameters:
        - name: server
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Server config not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-required-permissions:
        - admin

  # ── Discord Bot ───────────────────────────────────────
...
  /discord-bot/seeding/config:
    get:
      tags: [Discord Bot]
      summary: Get seeding config
      responses:
        "200":
          description: Seeding configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/SeedingConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Config not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-required-permissions:
        - view:discord-bot
        - manage:discord-bot
    patch:
      tags: [Discord Bot]
      summary: Update seeding config
      description: Partial update; omitted fields are left unchanged (not clobbered with defaults).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
                seedThreshold:
                  type: integer
                resetThreshold:
                  type: integer
                dailyTime:
                  type: string
                  nullable: true
                timezone:
                  type: string
                  nullable: true
                serverName:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - manage:discord-bot
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /auth/sync
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NoContent:
      description: Success (no content)
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
...
    UpdatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            updated:
              type: boolean
              example: true

    # ── Pagination ──────────────────────────────────────
...
    ServerConfig:
      type: object
      properties:
        id:
          type: string
        server:
          type: string
        label:
          type: string
        sftpHost:
          type: string
          nullable: true
        sftpPort:
          type: integer
        sftpUser:
          type: string
          nullable: true
        sftpPass:
          type: string
          nullable: true
          description: Always masked as "********" in responses
        sftpPath:
          type: string
          nullable: true
        syncEnabled:
          type: boolean

    # ── SquadJS Config ──────────────────────────────────
    SquadJSPlugin:
      type: object
      required: [plugin, enabled]
      properties:
        plugin:
          type: string
          description: Plugin class name
        enabled:
          type: boolean
      additionalProperties: true

    # ── Discord Bot ─────────────────────────────────────
    SeedingConfig:
      type: object
      properties:
        id:
          type: integer
        enabled:
          type: boolean
        channelId:
          type: string
          nullable: true
        roleId:
          type: string
          nullable: true
        seedThreshold:
          type: integer
        resetThreshold:
          type: integer
        dailyTime:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        serverName:
          type: string
          nullable: true
...